Refactor point picking
